pacdiff: Reduce repetition in input loop
The input loop repeats the `ask` code twice. In a future commit, we will
be introducing another instance of the `ask` code. This makes it
possible to mistakenly make the code to get out of sync.

Refactor the loop so that the `ask` code happens once at the top of the
loop. This introduces one functional change in the "Invalid answer" case
but it also makes it more consistent with the remaining cases.

Signed-off-by: Denton Liu <liu.denton@gmail.com>
Signed-off-by: Daniel M. Capella <polyzen@archlinux.org>